refactor(tauri): split terminal view responsibilities
